Deleted old Begin Workout fragment
Changed the layout for recording a workout, list items in begin workout
are clickable.
