+++ app/src/main/java/teameleven/smartbells2/BeginWorkout.java
@@ -1,94 +1,122 @@
 package teameleven.smartbells2;
 
-import android.content.Intent;
 import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentTabHost;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.Button;
-import android.widget.ListView;
-
-import java.sql.SQLException;
-import java.util.ArrayList;
+import android.view.animation.DecelerateInterpolator;
+import android.widget.TabHost;
+import android.widget.TextView;
 
 import teameleven.smartbells2.businesslayer.localdatabase.DatabaseAdapter;
-import teameleven.smartbells2.create.CreateCustomSession;
-import teameleven.smartbells2.dashboardfragmenttabs.RecordWorkoutRoutine;
+import teameleven.smartbells2.dashboardfragmenttabs.MyRoutines_Fragment;
+import teameleven.smartbells2.dashboardfragmenttabs.PublicRoutines_Fragment;
 
 /**
- * Created by Jare on 2015-10-06.
+ * Created by Jare on 2015-11-19.
  */
-public class BeginWorkout extends Fragment implements View.OnClickListener, AdapterView.OnItemClickListener {
+public class BeginWorkout extends Fragment {
 
-    Button createCustomSession;
-    private int checkTabPage;
-    private DatabaseAdapter db;
+    private FragmentTabHost dashboardTabHost;
     //value to be passed to RecordWorkout Activity
     public static final String ITEM_NAME = DatabaseAdapter.ROUTINE_NAME;
+    private int checkTabPage;
+    private FloatingActionButton fab;
 
-    private ArrayList<String> list;
-    private ArrayList<String> plist;
-
-    @Override
+    //Create the view and setup the tab host to view
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
-        //Main view
-        View view = inflater.inflate(R.layout.tab_fragment_beginworkout, container, false);
-        //Custom Session Button; when clicked it will call onClick
-        createCustomSession = (Button) view.findViewById(R.id.custom_session);
-        createCustomSession.setOnClickListener(this);
+        View rootView = inflater.inflate(R.layout.tab_beginworkout_dashboard, container, false);
 
-        //Get Routine list and fill up the list view
-        ListView routinelist = (ListView) view.findViewById(R.id.myRoutinelistview);
-        ListView publiclist = (ListView) view.findViewById(R.id.publicRoutinelistview);
-        //Set On Item Click Listener; clicking on a list item will call onItemClick
-        routinelist.setOnItemClickListener(this);
-        publiclist.setOnItemClickListener(this);
+        dashboardTabHost = (FragmentTabHost) rootView.findViewById(android.R.id.tabhost);
+        dashboardTabHost.setup(getActivity(), getChildFragmentManager(), R.id.realtabcontent);
 
-        //Open Database
-        db = new DatabaseAdapter(getActivity());
-        try {
-            db.openLocalDatabase();
-            //insert more routines
-            //db.insertTESTRoutines();
-        } catch (SQLException e) {
-            e.printStackTrace();
+
+        //Show the FAB
+        fab = (FloatingActionButton) getActivity().findViewById(R.id.fab);
+        fab.animate().translationY(0).setInterpolator(new DecelerateInterpolator(2)).start();
+
+        //Creates a listener, and handles fragment manager to swap between list fragments
+        TabHost.OnTabChangeListener tabChangeListener = new TabHost.OnTabChangeListener() {
+            //Inner method to control tab functions
+            @Override
+            public void onTabChanged(String tabId) {
+
+                android.support.v4.app.FragmentManager fragManager = getActivity().getSupportFragmentManager();
+
+                MyRoutines_Fragment myroutines_fragment = (MyRoutines_Fragment) fragManager.findFragmentByTag("myroutines");
+                PublicRoutines_Fragment publicroutines_fragment = (PublicRoutines_Fragment) fragManager.findFragmentByTag("publicroutines");
+
+                android.support.v4.app.FragmentTransaction fragTransaction = fragManager.beginTransaction();
+
+                //Detach myroutines_fragment if it exists
+                if (myroutines_fragment != null) {
+                    fragTransaction.detach(myroutines_fragment);
+                }
+
+                //Detach publicroutines_fragment if it exists
+                if (publicroutines_fragment != null) {
+                    fragTransaction.detach(publicroutines_fragment);
+                }
+
+                //If the current tab is on myroutines show only private routines
+                if (tabId.equalsIgnoreCase("myroutines")) {
+                    setCheckTabPage(1);
+                    if (myroutines_fragment == null) {
+                        fragTransaction.add(R.id.realtabcontent, new MyRoutines_Fragment(), "myroutines");
+                    } else {
+                        fragTransaction.attach(myroutines_fragment);
+                    }
+                }
+
+                //If the current tab is on publicroutines show only public routines
+                if (tabId.equalsIgnoreCase("publicroutines")) {
+                    setCheckTabPage(2);
+                    if (publicroutines_fragment == null) {
+                        fragTransaction.add(R.id.realtabcontent, new PublicRoutines_Fragment(), "publicroutines");
+                    } else {
+                        fragTransaction.attach(publicroutines_fragment);
+                    }
+                }
+
+                fragTransaction.commit();
+
+            }
+        };
+
+
+        //Setting up the tab change listener
+
+        dashboardTabHost.setOnTabChangedListener(tabChangeListener);
+
+        //Setting up builders for each tabs
+
+        //Achievements Tab creation
+        dashboardTabHost.addTab(dashboardTabHost.newTabSpec("myroutines").setIndicator("My Routines"),MyRoutines_Fragment.class, null);
+
+        //Records Tab creation
+        dashboardTabHost.addTab(dashboardTabHost.newTabSpec("publicroutines").setIndicator("Public Routines"), PublicRoutines_Fragment.class, null);
+
+
+        //Adjusting the text within the tabs to properly show titles
+        for (int i = 0; i < 2; i++)
+        {
+            TextView x = (TextView) dashboardTabHost.getTabWidget().getChildAt(i).findViewById(android.R.id.title);
+            if (i == 2)
+            {
+                x.setTextSize(15);
+            } else {x.setTextSize(15);}
+
+
         }
-        //Two ArrayLists for each ListView
-        list = db.getMyRoutinesAsStrings(db.getUserIDForSession());
-        plist = db.getRoutinesAsStrings();
 
-        //close the database
-        db.closeLocalDatabase();
-                                                                      //The custom layout,     the textview id,    the list
-        routinelist.setAdapter(new ArrayAdapter<String>(getActivity(), R.layout.list_item_routine, R.id.routineNameValueDetail, list));
-        publiclist.setAdapter(new ArrayAdapter<String>(getActivity(), R.layout.list_item_routine,R.id.routineNameValueDetail, plist));
-
-        return view;
+        //Return the view
+        return rootView;
     }
 
-    @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.custom_session:
-                Intent intent = new Intent(getActivity(), CreateCustomSession.class);
-                startActivity(intent);
-        }
-    }
-
-
-    @Override
-    public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-        //Start an intent when a list item is clicked
-
-        Intent intent = new Intent(getActivity(), RecordWorkoutRoutine.class);
-        //When we start the new intent we want to pass the name of the Routine from the list
-        intent.putExtra(ITEM_NAME, list.get(position));
-        startActivity(intent);
-    }
 
     public int getCheckBWTabPage() {
         return checkTabPage;
