+++ app/src/main/java/teameleven/smartbells2/dashboardfragmenttabs/RecordWorkoutRoutine.java
@@ -4,9 +4,10 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
-import teameleven.smartbells2.BeginWorkout;
 import teameleven.smartbells2.R;
+import teameleven.smartbells2.BeginWorkout;
 
 /**
  * Created by Jare on 2015-11-03.
@@ -25,6 +26,11 @@ public class RecordWorkoutRoutine extends AppCompatActivity {
         nameValue = getIntent().getStringExtra(BeginWorkout.ITEM_NAME);
         nameView = (TextView) findViewById(R.id.nameTextValue);
         nameView.setText(nameValue);
+    }
+
+    //Save new workout
+    public void saveSession(View view) {
+        Toast.makeText(this, "Save a new Workout Session", Toast.LENGTH_LONG).show();
     }
 
     //Cancel - back to menu
