+++ app/src/main/res/layout/record_workout.xml
@@ -11,19 +11,23 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        android:layout_margin="2dp">
+        android:layout_margin="0dp"
+        android:background="@color/colorPrimary"
+        >
+
         <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/blue"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:textSize="34dp"
-            android:text="name"
             android:id="@+id/nameTextValue"
-            android:layout_margin="2dp"/>
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="name"
+            android:textAppearance="?android:attr/textAppearanceLarge"
+            android:textColor="@color/white"
+            android:textSize="34dp"
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:layout_margin="30dp"/>
 
     </LinearLayout>
-
 
     <!-- SET GROUPS TITLE -->
 
@@ -31,7 +35,8 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        android:layout_margin="0dp">
+        android:layout_margin="10dp">
+
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -40,86 +45,103 @@
             android:id="@+id/setgroupsTextView"
             android:layout_margin="0dp"/>
 
-    </LinearLayout>
-
     <!-- SET GROUPS -->
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_margin="2dp">
-
         <!-- EXERCISE -->
-        <TextView
-            android:id="@+id/exerciseTextValue"
-            android:layout_width="wrap_content"
+
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="exercisename"
-            android:layout_margin="2dp"
-            android:layout_weight=".5"/>
+            android:orientation="horizontal"
+            >
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:textAppearance="?android:attr/textAppearanceLarge"
+                android:text="Exercise_Name"
+                android:id="@+id/exerciseTextViewCustom"
+                android:layout_margin="2dp"
+                android:layout_weight="1"/>
+
+        </LinearLayout>
 
         <!-- RESISTANCE LEVEL -->
 
-        <TextView
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:text="Resistance Level:"
-            android:id="@+id/resistanceTextView"
-            android:layout_margin="1dp"
-            android:layout_weight=".2"
-            />
-        <EditText
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:inputType="number"
-            android:ems="10"
-            android:id="@+id/editResistanceTextCustom"
-            android:layout_weight=".1" />
+            android:orientation="horizontal">
 
-        <!-- Reps Per Set -->
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:text="Reps Per Set:"
-            android:id="@+id/repsTextView"
-            android:layout_margin="1dp"
-            android:layout_weight=".2"
-            />
-        <EditText
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:inputType="number"
-            android:ems="10"
-            android:id="@+id/editRepsTextCustom"
-            android:layout_weight=".1" />
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textAppearance="?android:attr/textAppearanceMedium"
+                android:text="Resistance:"
+                android:id="@+id/resistanceTextViewCustom"
+                android:layout_weight="1"/>
 
-        <!-- Sets -->
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:text="Number of Sets:"
-            android:id="@+id/setsTextView"
-            android:layout_margin="0dp"
-            android:layout_weight=".2"
-            />
-        <EditText
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:inputType="number"
-            android:ems="10"
-            android:id="@+id/editSetsTextCustom"
-            android:layout_weight=".1" />
+            <EditText
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:inputType="number"
+                android:ems="10"
+                android:id="@+id/editResistanceText"
+                android:layout_weight="1" />
+
+        </LinearLayout>
+
+        <!-- SETS -->
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textAppearance="?android:attr/textAppearanceMedium"
+                android:text="Sets"
+                android:id="@+id/setsTextViewCustom"
+                android:layout_margin="1dp"
+                android:layout_weight=".2"
+                />
+            <EditText
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:inputType="number"
+                android:ems="10"
+                android:id="@+id/editSetsTextCustom"
+                android:layout_weight=".1"
+                android:numeric="integer" />
+        </LinearLayout>
+
+        <!-- Reps -->
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:textAppearance="?android:attr/textAppearanceMedium"
+                android:text="Reps Per Set"
+                android:id="@+id/repsTextViewCustom"
+                android:layout_margin="1dp"
+                android:layout_weight="1.1"/>
+
+            <EditText
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:inputType="number"
+                android:ems="5"
+                android:id="@+id/editRepsTextCustom"
+                android:layout_weight=".5"
+                android:numeric="integer" />
+        </LinearLayout>
 
     </LinearLayout>
 
     <!-- Save and Cancel BUTTONS -->
     <LinearLayout
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
         android:layout_margin="2dp">
