+++ app/src/main/res/layout/workout_set_group.xml
@@ -35,7 +35,7 @@
             android:id="@+id/editResistanceText"
             android:layout_weight="1" />
 
-        </LinearLayout>
+    </LinearLayout>
 
         <!-- Reps Per Set -->
         <LinearLayout
